make virtual env creation much faster
We first create a base venv that is not to be deleted and can be re-used easily.
There we install all default packages.

Then, we create another, kernel-specific, venv without any additional packages.
Instead, we create a link that points to the packages installed in the base image.

The base image is intended to stay immutable. All newly installed packages end up
in the kernel-specific venv. While they are lost on restart, the packages inside the
base image are not.